Added Prac4_6

+++ Pratical set 3/prac3_1.java
@@ -1,18 +1,16 @@
-//author : Shrey lakhtaria
 import java.util.*;
 
-public class prac3_1 
-{
+//author : Shrey lakhtaria
+
+public class prac3_1 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
         System.out.println("Enter the amount in pounds: ");
-        int pound = sc.nextInt();
-        int rupee;
+        double pound = sc.nextDouble();
+        double rupee;
         rupee = pound * 100;
         System.out.println("Equivalent Rupees: " + rupee);
         sc.close();
-
     }
 }
-
